pyproject.toml
[build-system]
requires = ["setuptools>=69", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-tooter"
version = "0.1.0"
description = "A CLI tool that posts photos to Mastodon using embedded metadata."
authors = [
    { name = "Jason Dentler" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "Mastodon.py",
]

[project.scripts]
photo-tooter = "photo_tooter.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "build",
    "pre-commit",
]
